Fix output of check and in calls.
The "check" and "in" calls now return the correct output that the
Concourse JSON parser expects (a list).

Moreover this commit fixes a bug appearing when the check is called:
Concourse does not pass any additional arguments and thus the reading
of the "workspace" value must be proteced otherwise Concourse displays
an error on the check resource.